Basic Refactoring of landing page
